% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlywarning.R
\name{earlywarning}
\alias{earlywarning}
\title{Function to determine the prediction interval for new value in time series}
\usage{
earlywarning(
  accidents,
  exposition = NULL,
  from = NULL,
  until = NULL,
  n = 10000,
  pred.level = NULL,
  main = NULL,
  x_axis = NULL,
  max_y = NULL,
  orientation_x = NULL,
  add_exp = FALSE,
  lang = "en"
)
}
\arguments{
\item{accidents}{Either an R date/time or character vector with accident data. For character vectors, the following data formats are allowed '2014-04-22', '2014/04/22' respectively '22.4.2014'}

\item{exposition}{Optional data frame with exposition data. The first column is the time value, the second column the exposure. If the time value is a specific date (e.g. '22.4.2014'), this is considered as the start date of this exposure. If the time value is a year (format '2010') the exposure is taken for the whole year. Exposure values are extended until a new entry is available. If necessary, the first exposure value is extended back forwards. DEFAULT NULL}

\item{from}{From which date or year (1.1.) the time series should be considered. Optional. If not specified, the 1.1 from the year of the earliest accident is used.}

\item{until}{Until when date or year (31.12) the time series should be considered. Optional. If not specified, the 31.12 from the year of the latest accident is used.}

\item{n}{Number of simulations}

\item{pred.level}{Level of the prediction interval, if NULL (default) an interval with return periods is constructed}

\item{main}{Optional title for the plot}

\item{x_axis}{Optional vector with the values for the x-axis}

\item{max_y}{Optional maximum value for the y-axis}

\item{orientation_x}{Alignment of the labels of the x-axis; "v" for vertical, "h" for horizontal, by default horizontal alignment is selected for 8 years or less, above that a vertical.}

\item{add_exp}{Option to supplement the output plot with the exposure as an additional axis. Additionally, a plot of the exposure alone is produced. Only activ if exposure is given.}

\item{lang}{Language for output ("en", "fr", "de" or "it")}
}
\description{
Function to determine the prediction interval for new value in time series
}
\examples{
  ex1 <- earlywarning(accidents = example1_timeserie)
  print(ex1)
  plot(ex1)
  summary(ex1)
  summary(ex1$fit)
  ex2 <- earlywarning(accidents = example2_timeserie, lang = "de")
  ex2
  ex3 <- earlywarning(accidents = example3_timeserie, lang = "it")
  print(ex3)
  ex4 <- earlywarning(accidents = example4_timeserie, lang = "de")
  print(ex4)
  plot(ex4)
  summary(ex4)
  ex5 <- earlywarning(accidents = example1_timeserie, exposition=exposition_ex1)
  plot(ex5)
  ex6 <- earlywarning(accidents = example1_timeserie, exposition=exposition_ex2, add_exp = TRUE)
  print(ex6)
  plot(ex6$plot_exposition)
  ex7 <- earlywarning(accidents = example3_timeserie, exposition=exposition_ex3)
  summary(ex7)
}
