% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{effectiveness}
\alias{effectiveness}
\title{Function to evaluate the effectiveness of measures}
\usage{
effectiveness(
  accidents,
  measure_start,
  measure_end,
  exposition = NULL,
  from = NULL,
  until = NULL,
  main = NULL,
  x_axis = NULL,
  y_axis = NULL,
  max_y = NULL,
  KI_plot = TRUE,
  silent = TRUE,
  lang = "en"
)
}
\arguments{
\item{accidents}{vector with accident data. The following data formats are allowed '2014-04-22', '2014/04/22' respectively '22.4.2014'}

\item{measure_start}{Start of the implementation of measures (format '2014-04-22', '2014/04/22' respectively '22.4.2014')}

\item{measure_end}{End of measure implementation respectively first day after the measure (format '2014-04-22', '2014/04/22' respectively '22.4.2014')}

\item{exposition}{optinal Dataframe with expositions data. The first column is the time value, the second column the exposure. If the time value is a specific date (e.g. '22.4.2014'), this is considered as the start date of this exposure. If the time value is a year (format '2010') the exposure is taken for the whole year. Exposure values are extended until a new entry is available. If necessary, the first exposure value is extended back forwards. DEFAULT NULL}

\item{from}{From which date or year (1.1) the time series should be considered. Optional. If not specified, the 1.1 from the year of the earliest accident is used.}

\item{until}{Until when date or year (31.12) the time series should be considered. Optional. If not specified, the 31.12 from the year of the latest accident is used.}

\item{main}{optional title for the plot}

\item{x_axis}{optional vector with the values for the x-axis}

\item{y_axis}{optional vector with the values for the y-axis}

\item{max_y}{optional maximum value for the y-axis}

\item{KI_plot}{TRUE/FALSE if an additional illustration with the 95\% confidence interval for the measure effect should be produced (only of limited use for models without measure effect)}

\item{silent}{parameter to suppress error messages during model evaluation}

\item{lang}{language for output ("en", "fr", "de" or "it")}
}
\description{
Function to evaluate the effectiveness of measures
}
\details{
Traffic accidents before and after the implementation of a traffic measure are analyzed to evaluate the effect of the #' measure. Since accident are counting data, they are modeled using count regression, by default with a Poisson model. #' However, the model is tested for overdispersion and in case of significant overdispersion it is automatically switched #' to a Negative binomial model. For the situation analysis, six different model scenarios for the measure are evaluated: #' no effect, trend, effect of measures, measure effect and trend, trend effect, measures and trend effect. The best
model is displayed. The exposure can optionally be considered as an offset.
}
\examples{
  # example
  ex1 <- effectiveness(accidents = example_no_effect, measure_start = '1.1.2011', measure_end = '1.1.2011')
  print(ex1)
  # plot(ex1)
  # plot_ci(ex1)
  # summary(ex1)
  ex2 <- effectiveness(accidents = example_measure_effect, measure_start = '1.1.2011', measure_end = '1.1.2011')
  ex3 <- effectiveness(accidents = example_measure_effect_and_trend , measure_start = '1.1.2011', measure_end = '1.1.2011')
  ex4 <- effectiveness(accidents = example_no_effect, measure_start = '1.1.2011', measure_end = '1.1.2011', exposition = exposition_ex1)

}
