% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{plot.class_effectiveness}
\alias{plot.class_effectiveness}
\title{Function to plot class effectiveness}
\usage{
\method{plot}{class_effectiveness}(object)
}
\description{
Function to plot class effectiveness
}
