% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness_multiple.R
\name{plot.class_effectiveness_multi}
\alias{plot.class_effectiveness_multi}
\title{Function to plot class_effectiveness_multi}
\usage{
\method{plot}{class_effectiveness_multi}(object)
}
\description{
Function to plot class_effectiveness_multi
}
