% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{Function to plot all the single case of an evaluation of a multiple location effectiveness evaluation}
\usage{
plot_all(object)
}
\arguments{
\item{object}{output of effectiveness_multi-Function}
}
\description{
Function to plot all the single case of an evaluation of a multiple location effectiveness evaluation
}
\examples{
  multi1 <- effectiveness_multiple(accidents = list(example_no_effect, example_measure_effect),
  measure_start = c('1.1.2011', '1.1.2012'), measure_end = c('1.1.2011', '1.1.2012'))
  plot_all(multi1)

}
