% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{plot_ci}
\alias{plot_ci}
\title{Function to directly plot confidence interval of measure}
\usage{
plot_ci(object)
}
\arguments{
\item{object}{output of effectiveness-Funktion}
}
\description{
Function to directly plot confidence interval of measure
}
