% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{print.class_effectiveness}
\alias{print.class_effectiveness}
\title{Function to print class effectiveness}
\usage{
\method{print}{class_effectiveness}(object, plot = TRUE)
}
\description{
Function to print class effectiveness
}
