% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness_multiple.R
\name{print.class_effectiveness_multi}
\alias{print.class_effectiveness_multi}
\title{Function to print class_effectiveness_multi}
\usage{
\method{print}{class_effectiveness_multi}(object, plot = TRUE)
}
\description{
Function to print class_effectiveness_multi
}
