% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{summary.class_effectiveness}
\alias{summary.class_effectiveness}
\title{Function for summary of class effectiveness}
\usage{
\method{summary}{class_effectiveness}(object)
}
\description{
Function for summary of class effectiveness
}
