% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness_multiple.R
\name{summary.class_effectiveness_multi}
\alias{summary.class_effectiveness_multi}
\title{Function for summary of class effectiveness_multi}
\usage{
\method{summary}{class_effectiveness_multi}(object)
}
\description{
Function for summary of class effectiveness_multi
}
