% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlywarning.R
\name{earlywarning}
\alias{earlywarning}
\title{Function to determine the prediction interval for new value in time series}
\usage{
earlywarning(
  accidents,
  exposition = NULL,
  n = 1000,
  conf.level = 0.9,
  from = NULL,
  until = NULL,
  main = NULL,
  silent = TRUE,
  x_axis = NULL,
  max_y = NULL,
  lang = "en",
  v = 1
)
}
\arguments{
\item{accidents}{vector with accident data. The following data formats are allowed '2014-04-22', '2014/04/22' respectively '22.4.2014'}

\item{exposition}{optinal Dataframe with expositions data. The first column is the time value, the second column the exposure. If the time value is a specific date (e.g. '22.4.2014'), this is considered as the start date of this exposure. If the time value is a year (format '2010') the exposure is taken for the whole year. Exposure values are extended until a new entry is available. If necessary, the first exposure value is extended back forwards. DEFAULT NULL}

\item{n}{number of simulations}

\item{conf.level}{level of the prediction interval}

\item{from}{From which date or year (1.1) the time series should be considered. Optional. If not specified, the 1.1 from the year of the earliest accident is used.}

\item{until}{Until when date or year (31.12) the time series should be considered. Optional. If not specified, the 31.12 from the year of the latest accident is used.}

\item{main}{optional title for the plot}

\item{silent}{parameter to suppress error messages during model evaluation}

\item{x_axis}{optional vector with the values for the x-axis}

\item{max_y}{optional maximum value for the y-axis}

\item{lang}{language for output ("en", "fr", "de" or "it")}
}
\description{
Function to determine the prediction interval for new value in time series
}
\examples{
  # example
  earlywarning(accidents=example_timeserie)
}
