% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earlywarning.R
\name{earlywarning}
\alias{earlywarning}
\title{Function to determine the prediction interval for new value in time series}
\usage{
earlywarning(
  accidents,
  exposition = NULL,
  n = 1000,
  conf.level = 0.9,
  from = NULL,
  until = NULL,
  main = NULL,
  silent = TRUE,
  x_axis = NULL,
  max_y = NULL,
  lang = "en",
  v = 1
)
}
\arguments{
\item{accidents}{Vector with accident data}

\item{exposition}{Dataframe with Expositionsdata (DTV) DEFAULT NULL}

\item{n}{number of simulations}

\item{conf.level}{level of the prediction interval}

\item{from}{From which date or year (1.1) the time series should be considered. Optional. If not specified, the 1.1 from the year of the earliest accident is used.}

\item{until}{Until when date or year (31.12) the time series should be considered. Optional. If not specified, the 31.12 from the year of the latest accident is used.}

\item{main}{optional title for the plot}

\item{silent}{parameter to suppress error messages during model evaluation}

\item{x_axis}{optional vector with the values for the x-axis}

\item{max_y}{optional maximum value for the y-axis}

\item{lang}{language for output (en, fr, de or it)}
}
\description{
Function to determine the prediction interval for new value in time series
}
\examples{
  # example
  earlywarning(accidents=example_timeserie)
}
