% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.R
\name{effectiveness}
\alias{effectiveness}
\title{Evaluation of effectiveness of measures}
\usage{
effectiveness(
  accidents,
  measure_start,
  measure_end,
  exposition = NULL,
  from = NULL,
  until = NULL,
  main = NULL,
  x_axis = NULL,
  max_y = NULL,
  orientation_x = NULL,
  add_exp = FALSE,
  KI_plot = TRUE,
  lang = "en"
)
}
\arguments{
\item{accidents}{Either an R date/time or character vector with accident dates. For character vectors, the following date formats are allowed '2014-04-22', '2014/04/22' respectively '22.4.2014'.}

\item{measure_start}{At which date the implementation of the measure started (e.g. character '22.4.2014' or R date/time).}

\item{measure_end}{At which date the implementation of the measure was terminated (respectively first day after the measure finished). If there is a period of acclimatization until road users have become accustomed to the new traffic regime, this can or should be integrated here.}

\item{exposition}{Optional data frame with exposition data. The first column is the time value, the second column the exposure. If the time value is a specific date (e.g. '22.4.2014'), this is considered as the start date of this exposure. If the time value is a year (format '2010') the exposure is taken for the whole year. Exposure values are extended until a new entry is available. If necessary, the first exposure value is extended backwards. DEFAULT NULL.}

\item{from}{From which date or year (1.1.) the time series should be considered. Optional. If not specified, the 1.1 from the year of the earliest accident is used.}

\item{until}{Until what date or year (31.12) the time series should be considered. Optional. If not specified, the 31.12 from the year of the latest accident is used.}

\item{main}{Optional title for the plot.}

\item{x_axis}{Optional vector with the values for the x-axis.}

\item{max_y}{Optional maximum value for the y-axis.}

\item{orientation_x}{Alignment of the labels of the x-axis; "v" for vertical, "h" for horizontal, by default horizontal alignment is selected for 8 years or less.}

\item{add_exp}{Option to supplement the output plot with the exposure as an additional axis. Furthermore an additionally plot of the exposure alone is produced. Only active if exposure is stated.}

\item{KI_plot}{TRUE/FALSE if an additional illustration with the 95\% confidence interval for the measure effect is produced (only of limited use for models without measure effect).}

\item{lang}{Language for output ("en", "fr", "de" or "it").}
}
\value{
A specific R object (\code{class_effectiveness}) is generated as function output. The main object is the illustration with a graphical analysis of the measures. The function \code{print.class_effectiveness()} is used to extract the most important key figures of the analysis.
Specifically, the output contains a list of the following elements:
\item{\code{fit}}{Output of the selected counting regression model (Negative Binomial or Poisson family.}
\item{\code{modelname}}{Selected model.}
\item{\code{data}}{Prepared data that were used for the analysis.}
\item{\code{pvalue_measure}}{p-value of the positive measure effect, if it exists.}
\item{\code{pvalue_interaction}}{p-value of the interaction term, if it exists.}
\item{\code{test_overdisp}}{p-value of the deviance dispersion test.}
\item{\code{plot}}{Plot graphical analysis (ggplot-class).}
\item{\code{plot_KI}}{Additional illustration with the 95\% confidence interval for the measure effect (ggplot-class).}
\item{\code{conf_limit}}{Overlapping of the confidence intervals before and after the measure.}
\item{\code{lang}}{Selected language.}
\item{\code{plot_exposition}}{Addional plot of the exposition, if available (ggplot-class).}
}
\description{
Function to evaluate the effectiveness of measures to reduce traffic accidents.
}
\details{
Traffic accidents before and after the implementation of a traffic measure are analyzed to evaluate the effect of the measure. Since accident are counting data, they are modeled using count regression, by default with a Poisson model. However, the model is tested for overdispersion and in case of significant overdispersion it is automatically switched to a Negative Binomial model. For the situation analysis, six different model scenarios for the measure are evaluated: no effect, trend, effect of measures, measure effect and trend, trend effect, measures and trend effect. The best model is displayed. The exposure can optionally be considered as an offset. The measure effect evaluates the difference directly before and directly after the measure.
An important assumption in the analysis is, that the implementation of measures is independent of the actual number of accidents. Accident numbers are random variables that fluctuate. If a measure is taken due to a randomly increased number of accidents, this leads to an overestimation of the effect of the measure in the analysis, since in such cases a decrease in the number of accidents can be expected even without a measure (regression-to-the-mean phenomenon). This is particularly problematic for site-specific measures with small accident numbers. Ideally, as the period prior to the measure, only the period after the decision to implement the measure was taken should be considered. In practice, this is difficult because the time periods are too short.
}
\examples{
  ex1 <- effectiveness(accidents = example_no_effect, measure_start = '1.1.2011', measure_end = '1.1.2011')
  print(ex1)
  plot(ex1)
  plot_ci(ex1)
  summary(ex1)
  ex2 <- effectiveness(accidents = example_measure_effect, measure_start = '1.1.2012', measure_end = '1.1.2012')
  ex2
  ex3 <- effectiveness(accidents = example_measure_and_trend_effect, measure_start = '2011-01-01', measure_end = '2011-1-1')
  plot(ex3)
  ex4 <- effectiveness(accidents = example_measure_effect_and_trend, measure_start = '2012/01/01', measure_end = '2012/1/1')
  ex4
  ex5 <- effectiveness(accidents = example_trend, measure_start = '1.1.2013', measure_end = '1.1.2013')
  print(ex5)
  ex6 <- effectiveness(accidents = example_trend_effect, measure_start = '1.1.2011', measure_end = '1.1.2011')
  print(ex6)
  ex7 <- effectiveness(accidents = example_no_effect, measure_start = '1.1.2011', measure_end = '1.1.2011', exposition = exposition_ex1, lang = "de")
  summary(ex7)
  plot(ex7)
  ex8 <- effectiveness(accidents = example_measure_effect, measure_start = '1.1.2012', measure_end = '1.4.2012', exposition = exposition_ex2, add_exp = TRUE)
  plot(ex8)
  plot(ex8$plot_exposition)
}
\seealso{
\code{\link[STAAD:effectiveness_multiple]{effectiveness_multiple()}} for the joint analysis of a measure that was implemented at several locations.
}
