% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness_multiple.R
\name{effectiveness_multiple}
\alias{effectiveness_multiple}
\title{Function to evaluate the effectiveness of measures with data from multiple location}
\usage{
effectiveness_multiple(
  accidents,
  measure_start,
  measure_end,
  exposition = NULL,
  from = NULL,
  until = NULL,
  main = NULL,
  x_axis = NULL,
  max_y = NULL,
  orientation_x = NULL,
  add_exp = FALSE,
  KI_plot = FALSE,
  lang = "en"
)
}
\arguments{
\item{accidents}{A list with either an R date/time or character vectors with accident data for each location. For character vectors, the following data formats are allowed '2014-04-22', '2014/04/22' respectively '22.4.2014'}

\item{measure_start}{A vector with the dates (R date/time as well as character in '2014-04-22', '2014/04/22' or '22.4.2014 format) of the implementation of the measure started for each location. The vector must either be the same length as the number of elements in the accident-list or contain only one value. If only one value is given, it is used for all locations.}

\item{measure_end}{A vector with the dates the implementation of the measure was finished for each location (first day after the measure). The vector must either be the same length as the number of elements in the accident-list or contain only one value. If only one value is given, it is used for all locations.}

\item{exposition}{optional list with data frame with exposition data. The first column is the time value, the second column the exposure. If the time value is a specific date (e.g. '22.4.2014'), this is considered as the start date of this exposure. If the time value is a year (format '2010') the exposure is taken for the whole year. Exposure values are extended until a new entry is available. If necessary, the first exposure value is extended back forwards. DEFAULT NULL}

\item{from}{Vector with dates or year (1.1.) from which the time series should be considered. Optional. If not specified, the 1.1 from the year of the earliest accident is used. Must be specified for none or all locations.  If only one value is given, it is used for all locations.}

\item{until}{Vector with dates or year (31.12) until which the time series should be considered. Optional. If not specified, the 31.12 from the year of the latest accident is used. Must be specified for none or all locations.  If only one value is given, it is used for all locations.}

\item{main}{optional title for the plot}

\item{max_y}{optional maximum value for the y-axis}

\item{orientation_x}{Alignment of the labels of the x-axis; "v" for vertical, "h" for horizontal, by default horizontal alignment is selected for 8 years or less, above that a vertical}

\item{add_exp}{Option to supplement the output plot with the exposure as an additional axis. Additionally, a plot of the exposure alone is produced. Only activ if exposure is given.}

\item{KI_plot}{TRUE/FALSE if an additional illustration with the 95\% confidence interval for the measure effect should be produced (only of limited use for models without measure effect)}

\item{lang}{language for output ("en", "fr", "de" or "it")}
}
\description{
Function to evaluate the effectiveness of measures with data from multiple location
}
\details{
Traffic accidents before and after the implementation of a traffic measure are analyzed to evaluate the effect of the #' measure. Since accident are counting data, they are modeled using count regression, by default with a Poisson model. #' However, the model is tested for overdispersion and in case of significant overdispersion it is automatically switched #' to a Negative binomial model. For the situation analysis, six different model scenarios for the measure are evaluated: #' no effect, trend, effect of measures, measure effect and trend, trend effect, measures and trend effect. The best
model is displayed. The exposure can optionally be considered as an offset.
}
\examples{
  multi1 <- effectiveness_multiple(accidents = list(example_no_effect, example_measure_effect),
  measure_start = c('1.1.2011', '1.1.2012'), measure_end = c('1.1.2011', '1.1.2012'))
  print(multi1)
  multi2 <- effectiveness_multiple(accidents = list(example_no_effect, example_measure_effect),
  measure_start = c('1.1.2011', '1.1.2012'), measure_end = c('1.1.2011', '1.1.2012'), exposition = exposition_ex2)
  print(multi2)
}
