% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ci.R
\name{plot_ci}
\alias{plot_ci}
\title{Function to directly plot confidence interval of measure}
\usage{
plot_ci(object)
}
\arguments{
\item{object}{output of effectiveness-Function}
}
\description{
Function to directly plot confidence interval of measure
}
\examples{
  # example
  ex1 <- effectiveness(accidents = example_no_effect, measure_start = '1.1.2011', measure_end = '1.1.2011')
  # plot_ci(ex)
}
