% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesanalysis.R
\name{timeseriesanalysis}
\alias{timeseriesanalysis}
\title{Trend evaluation in accident time series}
\usage{
timeseriesanalysis(
  accidents,
  exposition = NULL,
  from = NULL,
  until = NULL,
  pearson_line = TRUE,
  show_outliers = FALSE,
  main = NULL,
  max_y = NULL,
  min_y = NULL,
  x_axis = NULL,
  orientation_x = NULL,
  add_exp = FALSE,
  lang = "en"
)
}
\arguments{
\item{accidents}{Either an R date/time or character vector with accident dates. For character vectors, only the following date formats are allowed: '2014-04-22', '2014/04/22', '22.4.2014'.}

\item{exposition}{Optional data frame with exposition data. The first column is the time value, the second column the exposure. If the time value is a specific date (e.g. '22.4.2014'), this is considered as the start date of this exposure. If the time value is a year (format '2010') the exposure is taken for the whole year. Exposure values are extended until a new entry is available. If necessary, the first exposure value is extended backwards. DEFAULT NULL.}

\item{from}{From which date or year (1.1.) the time series should be considered. Optional. If not specified, the 1.1 from the year of the earliest accident is used.}

\item{until}{Until what date or year (31.12) the time series should be considered. Optional. If not specified, the 31.12 from the year of the latest accident is used.}

\item{pearson_line}{TRUE/FALSE, indicating if the line for Pearson residual equal 2 should be drawn or not.}

\item{show_outliers}{FALSE/TRUE, indicating if outliers with Pearson residual greater than 2 should be highlighted in color or not.}

\item{main}{Optional title for the plot.}

\item{max_y}{Optional maximum value for the y-axis.}

\item{min_y}{Optional minimum value for the y-axis, defaults to 0.}

\item{x_axis}{Optional, points at which tick-marks are to be drawn.}

\item{orientation_x}{Alignment of the labels of the x-axis; "v" for vertical, "h" for horizontal, by default horizontal alignment is selected for 8 years or less.}

\item{add_exp}{Option to supplement the output plot with the exposure as an additional axis. Furthermore an additional plot of the exposure alone is produced. Only active if exposure is available.}

\item{lang}{Language for output ("en", "fr", "de" or "it"), defaults to "en".}
}
\value{
A specific R object (\code{class_timeseriesanalyis}) is generated as function output. The main object is the plot with a graphical analysis of the trend in the time series. The function \code{print.class_timeseriesanalyis()} is used to extract the most important key figures of the analysis.
Specifically, the output contains a list of the following elements:
\item{\code{fit}}{Output of the count regression model (Poisson or Negative Binomial family.}
\item{\code{data}}{Prepared data that were used for the analysis.}
\item{\code{trend}}{Annual trend}
\item{\code{pvalue_trend}}{p-value for the null hypothesis that the trend equals 0.}
\item{\code{test_overdisp}}{p-value of the deviance dispersion test.}
\item{\code{plot}}{Plot graphical analysis (ggplot-class).}
\item{\code{lang}}{Selected language.}
\item{\code{plot_exposition}}{Addional plot of the exposition, if available (ggplot-class).}
}
\description{
Function to evaluate the trend in a traffic accident time series.
}
\details{
The function provides methodology for analyzing general trends in a time series of accident count data. It is based on the Poisson model. However, the fit is tested for overdispersion and in case of significant overdispersion it is automatically replaced by a Negative Binomial regression. The model enables the user to quantify observed trends and assess their statistical significance (reliability of results). The time resolution for accident periods is fixed at one year to avoid potential nuisance by seasonality. However, the start date for the analysis can be freely selected.
}
\examples{
  ex1 <- timeseriesanalysis(accidents = example1_timeserie)
  print(ex1)
  plot(ex1)
  summary(ex1)
  summary(ex1$fit)
  ex2 <- timeseriesanalysis(accidents = example2_timeserie, lang = "de")
  ex2
  ex3 <- timeseriesanalysis(accidents = example3_timeserie, lang = "it")
  print(ex3)
  ex4 <- timeseriesanalysis(accidents = example4_timeserie, lang = "de")
  print(ex4)
  plot(ex4)
  summary(ex4)
  ex5 <- timeseriesanalysis(accidents = example1_timeserie, exposition=exposition_ex1)
  plot(ex5)
  ex6 <- timeseriesanalysis(accidents = example1_timeserie, exposition=exposition_ex2)
  print(ex6)
  ex7 <- timeseriesanalysis(accidents = example1_timeserie, exposition=exposition_ex3, add_exp = TRUE)
  summary(ex7)
  plot(ex7)
  plot(ex7$plot_exposition)
}
\seealso{
\code{\link[STAAD:earlywarning]{earlywarning()}} function for early warning, which alerts the user if the current annual accident incidence is significantly higher than in previous years.
}
